+++ app/src/main/res/layout/activity_login_page.xml
@@ -4,71 +4,102 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".LoginPage">
+    android:orientation="vertical"
+    tools:context=".LoginActivity">
+
+    <TextView
+        android:id="@+id/efficiency"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_below="@+id/plan"
+        android:layout_alignStart="@+id/plan"
+        android:layout_marginStart="0dp"
+        android:layout_marginTop="-3dp"
+        android:text="@string/efficiency"
+        android:textAppearance="@style/TextAppearance.AppCompat.Display1"
+        android:textColor="#3e7a43" />
+
+    <TextView
+        android:id="@+id/plan"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_below="@+id/travel"
+        android:layout_alignStart="@+id/travel"
+        android:layout_alignParentEnd="true"
+        android:layout_marginStart="0dp"
+        android:layout_marginTop="-3dp"
+        android:layout_marginEnd="171dp"
+        android:text="@string/plan"
+        android:textAppearance="@style/TextAppearance.AppCompat.Display1"
+        android:textColor="#4eb274" />
 
     <Button
         android:id="@+id/signUpButton"
-        android:layout_width="145dp"
-        android:layout_height="63dp"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_alignParentEnd="true"
-        android:layout_marginStart="134dp"
-        android:layout_marginTop="479dp"
-        android:layout_marginEnd="132dp"
-        android:text="@string/Sign_up" />
+        android:background="@android:color/transparent"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center"
+        android:layout_below="@+id/efficiency"
+        android:layout_alignStart="@+id/travel"
+        android:layout_marginTop="240dp"
+        android:text="@string/signup" />
 
     <Button
         android:id="@+id/loginButton"
+        android:background="@android:color/transparent"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_alignParentEnd="true"
-        android:layout_marginStart="162dp"
-        android:layout_marginTop="376dp"
-        android:layout_marginEnd="161dp"
-        android:text="@string/Login" />
+        android:layout_below="@+id/efficiency"
+        android:layout_alignEnd="@+id/login_password"
+        android:layout_alignBottom="@+id/signUpButton"
+        android:layout_gravity="center"
+        android:layout_marginStart="33dp"
+        android:layout_marginTop="240dp"
+        android:layout_marginEnd="9dp"
+        android:layout_marginBottom="0dp"
+        android:layout_toEndOf="@+id/signUpButton"
+        android:text="@string/signin" />
 
     <EditText
         android:id="@+id/login_email"
-        android:layout_width="match_parent"
+        android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_alignBottom="@+id/login_password"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_alignParentEnd="true"
-        android:layout_marginStart="0dp"
-        android:layout_marginTop="196dp"
-        android:layout_marginEnd="0dp"
-        android:layout_marginBottom="56dp"
-        android:hint="@string/Email"
+        android:layout_below="@+id/efficiency"
+        android:layout_alignStart="@+id/efficiency"
+        android:layout_alignEnd="@+id/efficiency"
+        android:layout_marginStart="1dp"
+        android:layout_marginTop="93dp"
+        android:layout_marginEnd="-68dp"
+        android:hint="@string/username"
         android:inputType="textEmailAddress" />
 
 
     <EditText
         android:id="@+id/login_password"
-        android:layout_width="match_parent"
+        android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="0dp"
-        android:layout_marginTop="262dp"
-        android:hint="@string/Password"
+        android:layout_below="@+id/efficiency"
+        android:layout_alignStart="@+id/login_email"
+        android:layout_alignEnd="@+id/login_email"
+        android:layout_marginStart="-1dp"
+        android:layout_marginTop="142dp"
+        android:layout_marginEnd="0dp"
+        android:hint="@string/password"
         android:inputType="textPassword" />
 
     <TextView
-        android:id="@+id/Logo"
-        android:layout_width="110dp"
-        android:layout_height="100dp"
+        android:id="@+id/travel"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
         android:layout_alignParentStart="true"
         android:layout_alignParentTop="true"
         android:layout_alignParentEnd="true"
-        android:layout_marginStart="137dp"
-        android:layout_marginTop="59dp"
-        android:layout_marginEnd="137dp"
-        android:text="@string/Logo"
-        android:textAppearance="@style/TextAppearance.AppCompat.Display2" />
+        android:layout_marginStart="90dp"
+        android:layout_marginTop="89dp"
+        android:layout_marginEnd="170dp"
+        android:text="@string/travel"
+        android:textAppearance="@style/TextAppearance.AppCompat.Display1"
+        android:textColor="#81c487" />
 
 
 </RelativeLayout>
