+++ app/src/main/res/layout/activity_sign_up_page.xml
@@ -4,67 +4,106 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SignUpPage">
+    android:orientation="vertical"
+    tools:context=".SignUpActivity">
+
+    <EditText
+        android:id="@+id/last_name"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_below="@+id/sign_up_title"
+        android:layout_alignStart="@+id/sign_up_title"
+        android:layout_alignEnd="@+id/sign_up_title"
+        android:layout_alignBottom="@+id/login_password"
+        android:layout_marginStart="39dp"
+        android:layout_marginTop="83dp"
+        android:layout_marginEnd="26dp"
+        android:layout_marginBottom="56dp"
+        android:hint="@string/last_name"
+        android:inputType="textEmailAddress" />
+
+    <EditText
+        android:id="@+id/first_name"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_alignStart="@+id/sign_up_title"
+        android:layout_alignEnd="@+id/sign_up_title"
+        android:layout_alignBottom="@+id/login_password"
+        android:layout_alignParentTop="true"
+        android:layout_marginStart="39dp"
+        android:layout_marginTop="148dp"
+        android:layout_marginEnd="27dp"
+        android:layout_marginBottom="56dp"
+        android:hint="@string/first_name"
+        android:inputType="textEmailAddress" />
 
     <EditText
         android:id="@+id/signup_reenter"
-        android:layout_width="match_parent"
+        android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="0dp"
-        android:layout_marginTop="325dp"
-        android:hint="@string/Re_enter_password"
+        android:layout_below="@+id/sign_up_title"
+        android:layout_alignStart="@+id/sign_up_title"
+        android:layout_alignEnd="@+id/sign_up_title"
+        android:layout_marginStart="39dp"
+        android:layout_marginTop="247dp"
+        android:layout_marginEnd="30dp"
+        android:hint="@string/confirm"
         android:inputType="textPassword" />
 
     <EditText
         android:id="@+id/signup_password"
-        android:layout_width="match_parent"
+        android:layout_width="210dp"
         android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="0dp"
-        android:layout_marginTop="262dp"
-        android:hint="@string/Password"
+        android:layout_below="@+id/sign_up_title"
+        android:layout_alignStart="@+id/sign_up_title"
+        android:layout_alignEnd="@+id/sign_up_title"
+        android:layout_marginStart="39dp"
+        android:layout_marginTop="193dp"
+        android:layout_marginEnd="27dp"
+        android:hint="@string/password"
         android:inputType="textPassword" />
 
     <EditText
         android:id="@+id/signup_email"
-        android:layout_width="match_parent"
+        android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:layout_below="@+id/sign_up_title"
+        android:layout_alignStart="@+id/sign_up_title"
+        android:layout_alignEnd="@+id/sign_up_title"
         android:layout_alignBottom="@+id/login_password"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_alignParentEnd="true"
-        android:layout_marginStart="0dp"
-        android:layout_marginTop="196dp"
-        android:layout_marginEnd="0dp"
+        android:layout_marginStart="39dp"
+        android:layout_marginTop="137dp"
+        android:layout_marginEnd="26dp"
         android:layout_marginBottom="56dp"
-        android:hint="@string/Email"
+        android:hint="@string/email"
         android:inputType="textEmailAddress" />
 
     <TextView
-        android:id="@+id/logo"
-        android:layout_width="107dp"
-        android:layout_height="102dp"
+        android:id="@+id/sign_up_title"
+        android:layout_width="276dp"
+        android:layout_height="wrap_content"
         android:layout_alignParentStart="true"
         android:layout_alignParentTop="true"
         android:layout_alignParentEnd="true"
-        android:layout_marginStart="139dp"
-        android:layout_marginTop="63dp"
-        android:layout_marginEnd="138dp"
-        android:text="@string/Logo"
-        android:textAppearance="@style/TextAppearance.AppCompat.Display2" />
+        android:layout_centerHorizontal="true"
+        android:layout_marginStart="55dp"
+        android:layout_marginTop="34dp"
+        android:layout_marginEnd="53dp"
+        android:text="@string/sign_up_title"
+        android:textAppearance="@style/TextAppearance.AppCompat.Display1"
+        android:textColor="#81c487" />
 
     <Button
         android:id="@+id/signUpButton"
-        android:layout_width="145dp"
-        android:layout_height="63dp"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
         android:layout_alignParentStart="true"
         android:layout_alignParentTop="true"
         android:layout_alignParentEnd="true"
-        android:layout_marginStart="135dp"
-        android:layout_marginTop="459dp"
-        android:layout_marginEnd="131dp"
-        android:text="@string/Sign_up" />
+        android:layout_centerHorizontal="true"
+        android:layout_marginStart="91dp"
+        android:layout_marginTop="432dp"
+        android:layout_marginEnd="108dp"
+        android:background="@android:color/transparent"
+        android:text="@string/sign_up_button" />
 </RelativeLayout>
