+++ app/src/main/res/values/strings.xml
@@ -7,11 +7,19 @@
     <string name="nav_header_desc">Navigation header</string>
     <string name="action_settings">Settings</string>
     <string name="hello_world">hello world</string>
-
-    <string name="Email">Enter Your Email</string>
-    <string name="Password">Enter Your Password</string>
-    <string name="Re_enter_password">Re-enter Your password</string>
-    <string name="Sign_up">Sign Up</string>
-    <string name="Login">Login</string>
-    <string name="Logo">Logo</string>
+    <string name="travel">Travel</string>
+    <string name="plan">Plan</string>
+    <string name="efficiency">Efficiency</string>
+    <string name="signin">Sign In</string>
+    <string name="signup">Sign Up</string>
+    <string name="username">Username</string>
+    <string name="email">Email</string>
+    <string name="first_name">First Name</string>
+    <string name="last_name">Last Name</string>
+    <string name="password">Password</string>
+    <string name="confirm">Confirm Password</string>
+    <string name="sign_up_button">Start Your Journey!</string>
+    <string name="sign_up_title">...Tell us some basic information</string>
+    <string name="login">Login</string>
+   
 </resources>
